+++ wingchun/td/TDEngineCoinmex.h
@@ -124,7 +124,7 @@ private:
 
     ThreadPtr rest_thread;
     uint64_t last_rest_get_ts = 0;
-    int rest_get_interval_ms = 500;
+    uint64_t rest_get_interval_ms = 500;
 
     std::mutex* mutex_order_and_trade = nullptr;
